backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.29.3
moviepy==1.0.3
pytest==7.4.3
httpx==0.25.1
python-magic==0.4.27
pydantic-settings==2.1.0
tenacity==8.2.3
slowapi==0.1.8 